+++ app/Routes.js
@@ -1,15 +1,15 @@
 /* eslint flowtype-errors/show-errors: 0 */
 import React from 'react'
 import { Switch, Route } from 'react-router'
-import { HOME, SCHOOL } from './constants/routes.json'
+import { HOME, SCHOOLINFO } from './constants/routes.json'
 import App from './containers/App'
 import Home from './containers/Home'
-import School from './components/School'
+import SchoolInfo from './components/SchoolInfo'
 
 export default () => (
   <App>
     <Switch>
-      <Route path={SCHOOL} component={School} />
+      <Route path={SCHOOLINFO} component={SchoolInfo} />
       <Route path={HOME} component={Home} />
     </Switch>
   </App>
