+++ app/constants/routes.json
@@ -1,4 +1,4 @@
 {
   "HOME": "/",
-  "SCHOOL": "/school"
+  "SCHOOLINFO": "/SchoolInfo"
 }
