+++ app/reducers/schoolReducer.js
@@ -2,13 +2,15 @@
 import { HANDLE_SCHOOL_DATA } from '../actions/actionTypes'
 import type { Action } from './types'
 
+const _ = require('lodash')
+
 const initialLoadState = {}
 
 const handleSchoolReducer = (state = initialLoadState, action) => {
   switch (action.type) {
     case HANDLE_SCHOOL_DATA:
       const { title, schoolstate, state, country, zip, year } = action.payload
-      return Object.assign({}, state, {
+      return _.assign({}, state, {
         title,
         street,
         schoolstate,
@@ -16,7 +18,7 @@ const handleSchoolReducer = (state = initialLoadState, action) => {
         year
       })
     default:
-      return Object.assign({}, initialLoadState)
+      return state
   }
 }
