+++ package.json
@@ -210,6 +210,7 @@
     "devtron": "^1.4.0",
     "electron-debug": "^2.0.0",
     "history": "^4.7.2",
+    "lodash": "^4.17.11",
     "nedb": "^1.8.0",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
